Extract return type reflector to trait
