+++ src/Illuminate/Support/Traits/ReflectsClosures.php
@@ -6,6 +6,8 @@ use Closure;
 use Illuminate\Support\Collection;
 use Illuminate\Support\Reflector;
 use ReflectionFunction;
+use ReflectionIntersectionType;
+use ReflectionUnionType;
 use RuntimeException;
 
 trait ReflectsClosures
@@ -90,6 +92,35 @@ trait ReflectsClosures
 
                 return [$parameter->getName() => Reflector::getParameterClassName($parameter)];
             })
+            ->all();
+    }
+
+    /**
+     * Get the class names / types of the return type of the given Closure.
+     *
+     * @param  \Closure  $closure
+     * @return list<class-string>
+     *
+     * @throws \ReflectionException
+     */
+    protected function closureReturnTypes(Closure $closure)
+    {
+        $reflection = new ReflectionFunction($closure);
+
+        if ($reflection->getReturnType() === null ||
+            $reflection->getReturnType() instanceof ReflectionIntersectionType) {
+            return [];
+        }
+
+        $types = $reflection->getReturnType() instanceof ReflectionUnionType
+            ? $reflection->getReturnType()->getTypes()
+            : [$reflection->getReturnType()];
+
+        return (new Collection($types))
+            ->reject(fn ($type) => $type->isBuiltin())
+            ->reject(fn ($type) => in_array($type->getName(), ['static', 'self']))
+            ->map(fn ($type) => $type->getName())
+            ->values()
             ->all();
     }
 }
